Give recent api methods it's own controller
