+++ src/Server/Controllers/UsersController.cs
@@ -1,5 +1,8 @@
+using Application.Contracts;
 using Application.Exceptions;
 using Application.Users;
+using Blazorise.Extensions;
+using Domain.Contracts;
 using Domain.Users;
 
 using Microsoft.AspNetCore.Authorization;
