+++ tests/Client.Tests/Pages/RecentlyViewedTests.cs
@@ -13,7 +13,7 @@ public class RecentlyViewedTests : UITestFixture
     public void RecentlyViewedComponent_ShouldSayNothing_WhenThereAreNoRecentlyViewed()
     {
         // Arrange
-        MockHttp.When("/api/v1/contracts/recent/").RespondJson(Array.Empty<object>());
+        MockHttp.When("/api/v1/users//recents").RespondJson(Array.Empty<object>());
 
         // Act
         IRenderedComponent<RecentlyViewed> cut = Context.RenderComponent<RecentlyViewed>();
@@ -28,7 +28,7 @@ public class RecentlyViewedTests : UITestFixture
         // Arrange
         const string name = "SJ";
         var contract = new Contract() { Name = name, SupplierLogoImagePath = "/img/test" };
-        MockHttp.When("/api/v1/contracts/recent/").RespondJson(new[] { contract, });
+        MockHttp.When("/api/v1/users//recents").RespondJson(new[] { contract, });
 
         // Act
         IRenderedComponent<RecentlyViewed> cut = Context.RenderComponent<RecentlyViewed>();
