+++ tests/Server.Tests/Controllers/ContractsControllerTests.cs
@@ -10,16 +10,13 @@ public class ContractsControllerTests
 {
     private readonly ContractsController _cut;
     private readonly Mock<IContractService> _mockContracts;
-    private readonly Mock<IRecentContractService> _mockRecent;
 
     public ContractsControllerTests()
     {
         _mockContracts = new Mock<IContractService>();
-        _mockRecent = new Mock<IRecentContractService>();
         _cut = new ContractsController(
             Mock.Of<ILogger<ContractsController>>(),
-            _mockContracts.Object,
-            _mockRecent.Object);
+            _mockContracts.Object);
     }
 
     [Fact]
@@ -129,42 +126,5 @@ public class ContractsControllerTests
 
         // Assert
         contract.IsFavorite.Should().Be(false);
-    }
-
-    [Fact]
-    public void FetchRecent_ReturnsEmptyList_WhenUsernameIsEmpty()
-    {
-        // Act
-        IEnumerable<Contract> recentContracts = _cut.RecentContracts(string.Empty);
-
-        // Assert
-        recentContracts.Should().BeEmpty();
-    }
-
-    [Fact]
-    public void FetchRecent_DelegatesToService_WhenUsernameIsNotEmpty()
-    {
-        // Arrange
-        const string username = "123";
-
-        // Act
-        _cut.RecentContracts(username);
-
-        // Assert
-        _mockRecent.Verify(recent => recent.FetchRecentContracts(It.IsAny<string>()), Times.Once);
-    }
-
-    [Fact]
-    public void AddRecent_DelegatesToService()
-    {
-        // Arrange
-        const string username = "123";
-        var contract = new Contract();
-
-        // Act
-        _cut.AddRecent(username, contract);
-
-        // Assert
-        _mockRecent.Verify(recent => recent.Add(username, contract), Times.Once);
     }
 }
