+++ tests/Server.Tests/Controllers/UsersControllerTests.cs
@@ -1,6 +1,7 @@
+using Application.Contracts;
 using Application.Exceptions;
 using Application.Users;
-
+using Domain.Contracts;
 using Domain.Users;
 
 using Microsoft.AspNetCore.Mvc;
